Dodano CMake i wiele innych rzeczy ;)
