+++ .gitignore
@@ -1,0 +1,5 @@
+/build
+*.exe
+*.lib
+*.dll
+*.app
