+++ README.md
@@ -1,2 +1,5 @@
-# Spolszczenie-do-c-
+# Spolszczenie-do-c
 Spolszczenie składni c++, miłej zabawy !
+
+# Wymagania
+Program wymaga CMake w wersji wyższej niż 3.10
