+++ lib/src/spolszczenie.cpp
@@ -1,0 +1,1 @@
+#include "spolszczenie.h"
