+++ lib/src/spolszczenie.h
@@ -1,0 +1,101 @@
+#ifndef SPOLSZCZENIE_H
+#define SPOLSZCZENIE_H
+
+#define zawsze while (true)
+#define dopoki while
+#define zrob do
+
+#define start() int main()
+
+#define jezeli if
+#define inaczej else
+
+#define prawda true
+#define falsz false
+
+#define wyjscie std::cout
+#define wypisz printf
+
+#define nowa_linia "\n"
+
+#define calkowita int
+#define rzeczywista float
+#define dokladna double
+#define napis std::string
+#define znak char
+#define puste void
+
+#define przerwij break
+#define kontynuuj continue
+
+#define stala const
+#define staladoslowna constexpr
+
+#define nieoznaczona unsigned
+#define oznaczona signed
+#define krotka short int
+#define dluga long int
+#define bardzo_dluga long long int
+
+#define stworz new
+#define usun delete
+
+#define zwroc return
+
+#define klasa class
+#define struktura struct
+
+#define domyslna default
+
+#define idz_do goto
+
+#define wylicz enum
+
+#define zaprzyjaznij friend
+#define wlini inline
+#define plynna mutable
+#define przestrzennazw namespace
+
+#define prywatne private
+#define chronione protected
+#define publiczne public
+
+#define zarejestruj register
+
+#define rozmiar sizeof
+
+#define statyczna static
+
+#define wybor case
+#define decyzja switch
+
+#define dla for
+
+#define wzor template
+
+#define to this
+
+#define wyrzuc throw
+#define testuj try
+
+#define unia union
+
+#define uzyj using
+
+#define wirtualne virtual
+
+#define bitowe_przesuniecie_w_prawo <<
+#define bitowe_preszuniecie_w_lewo >>
+
+#define koniec std::endl
+
+#define dodaj +
+#define pomnoz *
+#define odejmij -
+#define podziel /
+#define komentarz //
+#define wskaznik *
+#define adres &
+#define wylaczjeslisiecoswpisalo cin.get()
+
+#endif // SPOLSZCZENIE_H
